chore(tracing): move zip and stackSession into packages/tracing/writer
Move the trace zip assembly (`zip`) and call-stack session bookkeeping
(`tracingStarted`, `traceDiscarded`, `addStackToTracingNoReply`) out of
`packages/playwright-core/src/server/localUtils.ts` and into
`packages/tracing/src/writer/`. The library functions take an
`AbortSignal` instead of a `Progress`; a tiny `race(signal, promise)`
helper in `packages/tracing/src/writer/race.ts` does the
abort-vs-promise race. The dispatcher passes `progress.signal` and
keeps the outer `progress.race(...)` so the eslint progress-rule
stays satisfied. HAR functions in `localUtils.ts` are left untouched.

The new package has no package.json; sources are bundled and
deps live in the root devDependencies. Per-file
\`node_modules/{fs,os,path,events,yauzl,yazl,mime}\` declarations
in \`packages/tracing/src/DEPS.list\` cover what writer/ needs.
