+++ packages/playwright-core/src/server/dispatchers/DEPS.list
@@ -3,5 +3,6 @@
 ../../protocol/
 @utils/**
 @isomorphic/**
+@tracing/**
 ../**
 node_modules/yaml
