+++ packages/playwright-core/src/server/localUtils.ts
@@ -15,124 +15,14 @@
  */
 
 import fs from 'fs';
-import os from 'os';
 import path from 'path';
 
-import * as yauzl from 'yauzl';
-import * as yazl from 'yazl';
-import { ManualPromise } from '@isomorphic/manualPromise';
-import { serializeClientSideCallMetadata } from '@tracing/reader/traceUtils';
-import { assert } from '@isomorphic/assert';
-import { calculateSha1 } from '@utils/crypto';
 import { ZipFile } from '@utils/zipFile';
-import { removeFolders } from '@utils/fileUtils';
 import { HarBackend } from './harBackend';
+
 import type * as channels from '@protocol/channels';
 import type * as har from '@tracing/format/har';
-import type EventEmitter from 'events';
 import type { Progress } from '@protocol/progress';
-
-
-export type StackSession = {
-  file: string;
-  writer: Promise<void>;
-  tmpDir: string | undefined;
-  callStacks: channels.ClientSideCallMetadata[];
-  live?: boolean;
-};
-
-export async function zip(progress: Progress, stackSessions: Map<string, StackSession>, params: channels.LocalUtilsZipParams): Promise<void> {
-  const promise = new ManualPromise<void>();
-  const zipFile = new yazl.ZipFile();
-  (zipFile as any as EventEmitter).on('error', error => promise.reject(error));
-
-  const addFile = (file: string, name: string) => {
-    try {
-      if (fs.statSync(file).isFile())
-        zipFile.addFile(file, name);
-    } catch (e) {
-    }
-  };
-
-  for (const entry of params.entries)
-    addFile(entry.value, entry.name);
-
-  // Add stacks and the sources.
-  const stackSession = params.stacksId ? stackSessions.get(params.stacksId) : undefined;
-  if (stackSession?.callStacks.length) {
-    await progress.race(stackSession.writer);
-    const buffer = Buffer.from(JSON.stringify(serializeClientSideCallMetadata(stackSession.callStacks)));
-    zipFile.addBuffer(buffer, 'trace.stacks');
-  }
-
-  // Collect sources from stacks.
-  if (params.includeSources) {
-    const sourceFiles = new Set<string>(params.additionalSources);
-    for (const { stack } of stackSession?.callStacks || []) {
-      if (!stack)
-        continue;
-      for (const { file } of stack)
-        sourceFiles.add(file);
-    }
-    for (const sourceFile of sourceFiles)
-      addFile(sourceFile, 'resources/src@' + calculateSha1(sourceFile) + '.txt');
-  }
-
-  if (params.mode === 'write') {
-    // New file, just compress the entries.
-    await progress.race(fs.promises.mkdir(path.dirname(params.zipFile), { recursive: true }));
-    zipFile.end(undefined, () => {
-      zipFile.outputStream.pipe(fs.createWriteStream(params.zipFile))
-          .on('close', () => promise.resolve())
-          .on('error', error => promise.reject(error));
-    });
-    await progress.race(promise);
-    await deleteStackSession(progress, stackSessions, params.stacksId);
-    return;
-  }
-
-  // File already exists. Repack and add new entries.
-  const tempFile = params.zipFile + '.tmp';
-  await progress.race(fs.promises.rename(params.zipFile, tempFile));
-
-  yauzl.open(tempFile, (err, inZipFile) => {
-    if (err) {
-      promise.reject(err);
-      return;
-    }
-    assert(inZipFile);
-    let pendingEntries = inZipFile.entryCount;
-    inZipFile.on('entry', entry => {
-      inZipFile.openReadStream(entry, (err, readStream) => {
-        if (err) {
-          promise.reject(err);
-          return;
-        }
-        zipFile.addReadStream(readStream!, entry.fileName);
-        if (--pendingEntries === 0) {
-          zipFile.end(undefined, () => {
-            zipFile.outputStream.pipe(fs.createWriteStream(params.zipFile)).on('close', () => {
-              fs.promises.unlink(tempFile).then(() => {
-                promise.resolve();
-              }).catch(error => promise.reject(error));
-            });
-          });
-        }
-      });
-    });
-  });
-  await progress.race(promise);
-  await deleteStackSession(progress, stackSessions, params.stacksId);
-}
-
-async function deleteStackSession(progress: Progress, stackSessions: Map<string, StackSession>, stacksId?: string) {
-  const session = stacksId ? stackSessions.get(stacksId) : undefined;
-  if (!session)
-    return;
-  stackSessions.delete(stacksId!);
-  if (session.tmpDir)
-    await progress.race(removeFolders([session.tmpDir]));
-}
 
 export async function harOpen(progress: Progress, harBackends: Map<string, HarBackend>, params: channels.LocalUtilsHarOpenParams): Promise<channels.LocalUtilsHarOpenResult> {
   let harBackend: HarBackend;
@@ -193,30 +83,5 @@ export async function harUnzip(progress: Progress, params: channels.LocalUtilsHa
     await progress.race(fs.promises.unlink(params.zipFile));
   } finally {
     zipFile.close();
-  }
-}
-
-export async function tracingStarted(progress: Progress, stackSessions: Map<string, StackSession>, params: channels.LocalUtilsTracingStartedParams): Promise<channels.LocalUtilsTracingStartedResult> {
-  let tmpDir = undefined;
-  if (!params.tracesDir)
-    tmpDir = await progress.race(fs.promises.mkdtemp(path.join(os.tmpdir(), 'playwright-tracing-')));
-  const traceStacksFile = path.join(params.tracesDir || tmpDir!, params.traceName + '.stacks');
-  stackSessions.set(traceStacksFile, { callStacks: [], file: traceStacksFile, writer: Promise.resolve(), tmpDir, live: params.live });
-  return { stacksId: traceStacksFile };
-}
-
-export async function traceDiscarded(progress: Progress, stackSessions: Map<string, StackSession>, params: channels.LocalUtilsTraceDiscardedParams): Promise<void> {
-  await deleteStackSession(progress, stackSessions, params.stacksId);
-}
-
-export function addStackToTracingNoReply(stackSessions: Map<string, StackSession>, params: channels.LocalUtilsAddStackToTracingNoReplyParams) {
-  for (const session of stackSessions.values()) {
-    session.callStacks.push(params.callData);
-    if (session.live) {
-      session.writer = session.writer.then(() => {
-        const buffer = Buffer.from(JSON.stringify(serializeClientSideCallMetadata(session.callStacks)));
-        return fs.promises.writeFile(session.file, buffer);
-      });
-    }
   }
 }
